requirements.txt
absl-py==1.0.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
biopython==1.79
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.3.0
chardet==5.1.0
charset-normalizer==3.1.0
chex==0.1.5
click==8.1.3
contextlib2==21.6.0
cycler==0.11.0
cycless==0.3.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dm-haiku==0.0.9
dm-tree==0.1.6
dnspython==2.3.0
docker-pycreds==0.4.0
einops==0.6.0
email-validator==1.3.1
entrypoints==0.4
etils==0.9.0
fair-esm==2.0.0
fastapi==0.95.0
fastjsonschema==2.16.3
flatbuffers==1.12
fonttools==4.38.0
gast==0.4.0
gemmi==0.6.6
GenIce2==2.1.7.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphstat==0.2.1
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
htmlmin==0.1.12
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
ImageHash==4.3.1
immutabledict==2.0.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmp==0.0.4
joblib==1.2.0
jsonschema==4.17.3
jupyter-console==6.6.3
jupyter-server==1.23.6
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
methodtools==0.4.7
mistune==2.0.5
mkl-fft==1.3.1
mkl-service==2.4.0
ml-collections==0.1.0
multimethod==1.9.1
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.3
notebook_shim==0.2.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openpyscad==0.5.0
optax==0.1.4
orjson==3.8.8
packaging==23.0
PairList==0.3.2
pandas==1.3.4
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
plotting==0.0.7
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
py3Dmol==2.0.1.post1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyCifRW==4.4.6
pydantic==1.10.7
Pygments==2.14.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.6
pytz==2023.2
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
rdkit==2023.3.2
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==2.3.1
setproctitle==1.3.3
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
statsmodels==0.13.5
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
TileCycles==0.1.5.6
tinycss2==1.2.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.7.1
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
validators==0.20.0
visions==0.7.5
wandb==0.17.0
watchfiles==0.18.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
widgetsnbextension==4.0.6
wirerope==0.4.7
wrapt==1.15.0
yaplotlib==0.1.2
ydata-profiling==4.1.1
zipp==3.15.0
